renaming oracle to oracle-dbm
