+++ pkg/collector/corechecks/oracle-dbm/common/utils.go
