+++ pkg/collector/corechecks/oracle-dbm/sql/setup.sql
